refactor: 클래스명 변경 및 폴더 변경
issue: #11
