+++ src/main/java/com/kestrel/weddingbookkeeper/api/wedding/dto/WeddingInsertDto.java
@@ -2,7 +2,7 @@ package com.kestrel.weddingbookkeeper.api.wedding.dto;
 
 import com.kestrel.weddingbookkeeper.api.member.vo.Gender;
 import com.kestrel.weddingbookkeeper.api.member.vo.Member;
-import com.kestrel.weddingbookkeeper.api.wedding.util.AuthorizationCodeGenerator;
+import com.kestrel.weddingbookkeeper.api.wedding.util.ValidationCodeGenerator;
 import java.time.LocalDateTime;
 import lombok.ToString;
 
@@ -23,8 +23,8 @@ public class WeddingInsertDto {
         this.groomId = member.getGender() == Gender.MALE ? member.getId() : null;
         this.brideId = member.getGender() == Gender.FEMALE ? member.getId() : null;
         this.qrImgUrl = qrImgUrl;
-        this.partnerCode = AuthorizationCodeGenerator.generateQrCode();
-        this.managerCode = AuthorizationCodeGenerator.generateQrCode();
+        this.partnerCode = ValidationCodeGenerator.generateQrCode();
+        this.managerCode = ValidationCodeGenerator.generateQrCode();
         this.groomName = member.getGender() == Gender.MALE ? member.getName() : null;
         this.brideName = member.getGender() == Gender.FEMALE ? member.getName() : null;
         this.location = weddingInfoRequestDto.getLocation();
