+++ src/main/java/com/kestrel/weddingbookkeeper/api/wedding/facade/WeddingFacade.java
@@ -1,4 +1,4 @@
-package com.kestrel.weddingbookkeeper.api.member.facade;
+package com.kestrel.weddingbookkeeper.api.wedding.facade;
 
 import com.kestrel.weddingbookkeeper.api.wedding.dto.WeddingInfoRequestDto;
