+++ src/main/java/com/kestrel/weddingbookkeeper/common/configuration/MyBatisConfiguration.java
@@ -1,4 +1,4 @@
-package com.kestrel.weddingbookkeeper.common.config;
+package com.kestrel.weddingbookkeeper.common.configuration;
 
 import org.apache.ibatis.type.JdbcType;
 import org.mybatis.spring.boot.autoconfigure.ConfigurationCustomizer;
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class MyBatisConfig {
+public class MyBatisConfiguration {
 
     @Bean
     public ConfigurationCustomizer mybatisConfigurationCustomizer() {
