+++ src/main/java/com/kestrel/weddingbookkeeper/external/storage/service/FileService.java
@@ -1,4 +1,4 @@
-package com.kestrel.weddingbookkeeper.external.storage;
+package com.kestrel.weddingbookkeeper.external.storage.service;
 
 import java.io.InputStream;
